backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.1
httpx==0.27.2
pydantic==2.11.5
pydantic[email]==2.11.5
python-dotenv==1.0.1
motor==3.6.0
pymongo==4.9.2
python-jose[cryptography]==3.3.0

# passlib 1.7.4 is only compatible with bcrypt 3.x — bcrypt 4.x removed __about__
# This is the root cause of "AttributeError: module 'bcrypt' has no attribute '__about__'"
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

fpdf2==2.7.9
python-pptx==1.0.2
markdown2==2.5.0

# ── Phase 5: RAG ──────────────────────────────────────────────────────────────
chromadb==0.6.3
sentence-transformers==3.3.1
pypdf==5.1.0
python-multipart==0.0.20
duckduckgo-search==7.3.2
